requirement.txt
grpcio==1.71.0
grpcio-tools==1.71.0
transformers==4.51.3
torch==2.4.0
kokoro
pydub
soundfile